% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{calcBiCc}
\alias{calcBiCc}
\title{Function for bivariate analysis of two case-control studies}
\usage{
calcBiCc(
  ncase1 = 1000,
  ncase2 = 1000,
  ncontrol1 = 1000,
  ncontrol2 = 1000,
  hsq1 = 0.5,
  hsq2 = 0.5,
  K1 = 0.1,
  K2 = 0.1,
  rg = 0.5,
  overlap = FALSE,
  alpha = 0.05,
  var_pi = 2e-05
)
}
\arguments{
\item{ncase1}{number of cases in first study}

\item{ncase2}{number of cases in second study}

\item{ncontrol1}{number of controls in first study}

\item{ncontrol2}{number of controls in second study}

\item{hsq1}{variance explained by all SNPs in first trait}

\item{hsq2}{variance explained by all SNPs in second trait}

\item{K1}{disease prevalence of first trait}

\item{K2}{disease prevalence of second trait}

\item{rg}{genetic correlation}

\item{overlap}{boolean indicating whether there is sample overlap between the two studies}

\item{alpha}{significance level}

\item{var_pi}{variance of the off-diagonal elements of the GRM}
}
\value{
list containing se (standard error), ncp (non-centrality parameter) and power
}
\description{
Function for bivariate analysis of two case-control studies
}
