% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{calcBiQtCc}
\alias{calcBiQtCc}
\title{Function for bivariate analysis of a quantitative trait and a binary trait (case-control study)}
\usage{
calcBiQtCc(
  n = 1000,
  ncase = 1000,
  ncontrol = 1000,
  hsq1 = 0.5,
  hsq2 = 0.5,
  K = 0.1,
  rg = 0.5,
  overlap = FALSE,
  alpha = 0.05,
  var_pi = 2e-05
)
}
\arguments{
\item{n}{sample size of quantitative trait}

\item{ncase}{number of cases in case-control study}

\item{ncontrol}{number of controls in case-control study}

\item{hsq1}{variance explained by all SNPs in first trait}

\item{hsq2}{variance explained by all SNPs in second trait}

\item{K}{disease prevalence of case-control trait}

\item{rg}{genetic correlation}

\item{overlap}{boolean indicating whether there is sample overlap between the two studies}

\item{alpha}{significance level}

\item{var_pi}{variance of the off-diagonal elements of the GRM}
}
\value{
list containing se (standard error), ncp (non-centrality parameter) and power
}
\description{
Function for bivariate analysis of a quantitative trait and a binary trait (case-control study)
}
