% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{calcUniCc}
\alias{calcUniCc}
\title{Calculate power to identify heritability of binary trait}
\usage{
calcUniCc(
  ncase = 1000,
  ncontrol = 1000,
  hsq = 0.5,
  K = 0.1,
  alpha = 0.05,
  var_pi = 2e-05
)
}
\arguments{
\item{ncase}{number of cases}

\item{ncontrol}{number of controls}

\item{hsq}{variance explained by all SNPs}

\item{K}{disease prevalence in population}

\item{alpha}{significance level}

\item{var_pi}{variance of the off-diagonal elements of the GRM}
}
\value{
list containing se (standard error), ncp (non-centrality parameter) and power
}
\description{
Calculate power to identify heritability of binary trait
}
