% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{munge_gwas_input}
\alias{munge_gwas_input}
\title{From https://danielroelfs.com/blog/how-i-create-manhattan-plots-using-ggplot/}
\usage{
munge_gwas_input(gwas_dat)
}
\description{
From https://danielroelfs.com/blog/how-i-create-manhattan-plots-using-ggplot/
}
