% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{obs_scale_to_liab_scale}
\alias{obs_scale_to_liab_scale}
\title{Transform heritability (h2) estimates from the observed scale to the liability scale}
\usage{
obs_scale_to_liab_scale(h2.obs, P, K)
}
\arguments{
\item{h2.obs}{estimated h2 on the observed scale}

\item{P}{prevalence of disease}

\item{K}{case proportion in study}
}
\description{
Transform heritability (h2) estimates from the observed scale to the liability scale
}
