% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2.R
\name{var_vg_func}
\alias{var_vg_func}
\title{NB: code from the GCTA website: https://yanglab.westlake.edu.cn/software/gcta/#GREMLpowercalculator}
\usage{
var_vg_func(N, var_pi = 2e-05)
}
\description{
NB: code from the GCTA website: https://yanglab.westlake.edu.cn/software/gcta/#GREMLpowercalculator
}
